+++ src/main/java/com/fmi/domain/user/web/controller/UserController.java
@@ -14,27 +14,19 @@ import com.fmi.domain.user.response.UserMetaResponse;
 import com.fmi.domain.user.response.UserOtherPageResponse;
 import com.fmi.domain.user.response.UserProfileResponse;
 import com.fmi.domain.user.service.UserService;
-import com.fmi.domain.user.web.dto.AccountDeleteRequest;
-import com.fmi.domain.user.web.dto.PasswordChangeRequest;
-import com.fmi.domain.user.web.dto.PasswordVerifyRequest;
 import com.fmi.domain.user.web.dto.UserUpdateRequest;
 import com.fmi.global.apiPayload.ApiResponse;
-import com.fmi.security.CookieFactory;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseCookie;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
@@ -47,13 +39,6 @@ import org.springframework.web.multipart.MultipartFile;
 public class UserController {
 
     private final UserService userService;
-    private final CookieFactory cookieFactory;
-
-    @Value("${jwt.cookie.name:refresh_token}")
-    private String refreshCookieName;
-
-    @Value("${jwt.cookie.access-token-name:access_token}")
-    private String accessCookieName;
 
     @PostMapping("/uploads/images")
     @Operation(summary = "이미지 업로드", description = "여러 장의 이미지를 S3에 업로드하고 URL을 반환합니다. (JPEG, PNG 형식만 지원)")
@@ -359,158 +344,5 @@ public class UserController {
         boolean deleteProfileImage = request != null && request.isDeleteProfileImage();
         UserProfileResponse response = userService.updateMyProfile(email, request, profileImage, deleteProfileImage);
         return ApiResponse.onSuccess(response);
-    }
-
-    @PostMapping("/me/password/verify")
-    @Operation(summary = "현재 비밀번호 검증", description = "현재 비밀번호가 올바른지 검증합니다. 비밀번호 변경 전에 먼저 호출하여 비밀번호를 확인해야 합니다.")
-    @ApiResponses({
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "비밀번호 검증 성공"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "400",
-                description = "USER400-PASSWORD_INCORRECT: 현재 비밀번호가 일치하지 않습니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"USER400-PASSWORD_INCORRECT\", \"message\": \"현재 비밀번호가 일치하지 않습니다.\"}"))),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "404",
-                description = "USER404-NOT_FOUND: 존재하지 않는 회원입니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"USER404-NOT_FOUND\", \"message\": \"존재하지 않는 회원입니다.\"}")))
-    })
-    public ApiResponse<Void> verifyPassword(
-            @AuthenticationPrincipal UserDetails userDetails, @Valid @RequestBody PasswordVerifyRequest request) {
-        String email = userDetails.getUsername();
-        userService.verifyPasswordWithException(email, request);
-        return ApiResponse.onSuccess(null);
-    }
-
-    @PatchMapping("/me/password")
-    @Operation(
-            summary = "비밀번호 변경",
-            description =
-                    "새 비밀번호로 변경합니다. 비밀번호 검증은 별도 엔드포인트(/users/me/password/verify)에서 먼저 완료해야 합니다. 새 비밀번호는 8~16자, 대/소문자·숫자·특수문자를 포함해야 합니다.")
-    @ApiResponses({
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "비밀번호 변경 성공"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "400",
-                description = "USER400-PASSWORD_MISMATCH: 새 비밀번호와 확인이 일치하지 않습니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"USER400-PASSWORD_MISMATCH\", \"message\": \"새 비밀번호와 확인이 일치하지 않습니다.\"}"))),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "400",
-                description = "AUTH400-WEAK_PASSWORD: 비밀번호 규칙을 만족하지 않습니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"AUTH400-WEAK_PASSWORD\", \"message\": \"비밀번호 규칙을 만족하지 않습니다. 8~16자, 대/소문자·숫자·특수문자를 포함해야 합니다.\"}"))),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "404",
-                description = "USER404-NOT_FOUND: 존재하지 않는 회원입니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"USER404-NOT_FOUND\", \"message\": \"존재하지 않는 회원입니다.\"}")))
-    })
-    public ApiResponse<Void> changePassword(
-            @AuthenticationPrincipal UserDetails userDetails, @Valid @RequestBody PasswordChangeRequest request) {
-        String email = userDetails.getUsername();
-        userService.changePassword(email, request);
-        return ApiResponse.onSuccess(null);
-    }
-
-    @DeleteMapping("/me")
-    @Operation(summary = "회원 탈퇴", description = """
-                    현재 로그인한 사용자의 계정을 소프트 삭제합니다.
-
-                    **인증:**
-                    - 일반 회원: 현재 비밀번호 입력 필수
-                    - 소셜 로그인 회원 (카카오 등): 비밀번호 불필요 (password 필드 생략 가능)
-
-                    **삭제 방식:**
-                    - 즉시 소프트 삭제: 계정은 삭제 표시되지만 30일간 데이터가 보관됩니다
-                    - 프로필 이미지는 즉시 S3에서 삭제됩니다
-                    - 30일 후 자동으로 완전 삭제(하드 삭제)됩니다
-
-                    **복구 및 재가입:**
-                    - 30일 이내에는 복구가 가능합니다 (관리자 문의)
-                    - 탈퇴 후 7일 이내에는 동일한 이메일로 재가입할 수 없습니다
-                    - 7일 경과 후에는 동일한 이메일로 재가입이 가능합니다
-
-                    **주의사항:**
-                    - 탈퇴 후에는 로그인 및 서비스 이용이 불가능합니다
-                    - 작성한 게시글과 댓글은 자동으로 삭제되지 않으며, 익명화 처리될 수 있습니다
-                    - 탈퇴 시 인증 토큰(쿠키)이 즉시 만료 처리됩니다
-
-                    **탈퇴 사유:**
-                    - 탈퇴 사유를 선택해야 합니다
-                    - reason이 OTHER인 경우 otherReason에 상세 사유를 입력할 수 있습니다
-                    """)
-    @ApiResponses({
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "회원 탈퇴 성공"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "400",
-                description = "USER400-PASSWORD_INCORRECT: 현재 비밀번호가 일치하지 않습니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"USER400-PASSWORD_INCORRECT\", \"message\": \"현재 비밀번호가 일치하지 않습니다.\"}"))),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "404",
-                description = "USER404-NOT_FOUND: 존재하지 않는 회원입니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"USER404-NOT_FOUND\", \"message\": \"존재하지 않는 회원입니다.\"}"))),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                responseCode = "500",
-                description = "FILE500-DELETE_IO: 파일을 삭제할 수 없습니다",
-                content =
-                        @Content(
-                                mediaType = "application/json",
-                                examples =
-                                        @ExampleObject(
-                                                value =
-                                                        "{\"isSuccess\": false, \"code\": \"FILE500-DELETE_IO\", \"message\": \"파일을 삭제할 수 없습니다.\"}")))
-    })
-    public ResponseEntity<ApiResponse<Void>> deleteAccount(
-            @AuthenticationPrincipal UserDetails userDetails,
-            @Valid @RequestBody AccountDeleteRequest request,
-            HttpServletRequest httpRequest) {
-        String email = userDetails.getUsername();
-        userService.deleteAccount(email, request);
-
-        ResponseCookie removeAccess = cookieFactory.expire(httpRequest, accessCookieName);
-        ResponseCookie removeRefresh = cookieFactory.expire(httpRequest, refreshCookieName);
-
-        return ResponseEntity.ok()
-                .header("Set-Cookie", removeAccess.toString())
-                .header("Set-Cookie", removeRefresh.toString())
-                .body(ApiResponse.onSuccess(null));
     }
 }
